Support Django 2.0+
With the release of the 2.0 version of Django, the `on_delete` kwarg is
required when establishing a relationship. The previously default
behavior of `models.CASCADE`, as this is the usual behavior in database
systems, was selected as the `on_delete` kwarg in the tests for this
library.

After updating the tests, the build was green.